Migrate Centreon web and widgets to rspack
