+++ centreon/package.json
@@ -8,10 +8,10 @@
     "widgets:build:dev": "(cd www/widgets && pnpm build:dev $WIDGETS)",
     "widgets:watch": "(cd www/widgets && pnpm watch $WIDGETS)",
     "widgets:build:analyze": "(cd www/widgets && pnpm build:analyze)",
-    "centreon:build": "webpack --mode production --config ./webpack.config.prod.js",
-    "centreon:build:analyze": "webpack --mode production --config ./webpack.config.analyze.js",
-    "centreon:build:dev": "cross-env WEBPACK_ENV=development webpack serve --mode development --config ./webpack.config.dev.js",
-    "centreon:watch": "cross-env WEBPACK_ENV=serve webpack serve --mode development --config ./webpack.config.dev.js",
+    "centreon:build": "rspack build -c ./rspack.config.prod.js",
+    "centreon:build:analyze": "rspack build -c ./rspack.config.prod.js --analyze",
+    "centreon:build:dev": "rspack build -m development -c ./rspack.config.dev.js",
+    "centreon:watch": "rspack serve -c ./rspack.config.dev.js",
     "build": "pnpm widgets:build && pnpm centreon:build",
     "build:dev": "pnpm widgets:build:dev && pnpm centreon:build:dev",
     "build:watch": "pnpm centreon:watch & pnpm widgets:watch",
@@ -38,6 +38,9 @@
     "@cypress/webpack-preprocessor": "^6.0.1",
     "@mui/utils": "^5.15.4",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.11",
+    "@rspack/cli": "^0.6.3",
+    "@rspack/core": "^0.6.3",
+    "@rspack/plugin-react-refresh": "^0.6.3",
     "@svgr/webpack": "^8.1.0",
     "@swc/core": "^1.3.105",
     "@swc/core-linux-x64-gnu": "^1.3.105",
