+++ centreon/packages/js-config/rspack/patch/dev.js
@@ -1,0 +1,12 @@
+module.exports = {
+  getDevConfiguration: () => ({
+    cache: true,
+    devtool: 'eval-cheap-module-source-map',
+    optimization: {
+      splitChunks: false
+    },
+    output: {
+      filename: '[name].js'
+    }
+  })
+};
