+++ centreon/packages/js-config/rspack/patch/module.js
@@ -1,5 +1,3 @@
-const { CleanWebpackPlugin } = require('clean-webpack-plugin');
-
 const WriteRemoteEntryNameToModuleFederation = require('../plugins/WriteRemoteEntryNameToModuleFederation');
 const TransformPreloadScript = require('../plugins/TransformPreloadScript');
 
@@ -9,11 +7,6 @@ module.exports = ({ outputPath, federatedComponentConfiguration }) => ({
     path: outputPath
   },
   plugins: [
-    new CleanWebpackPlugin({
-      cleanOnceBeforeBuildPatterns: [`${outputPath}/**/*.js`],
-      dangerouslyAllowCleanPatternsOutsideProject: true,
-      dry: false
-    }),
     new WriteRemoteEntryNameToModuleFederation(federatedComponentConfiguration),
     new TransformPreloadScript(federatedComponentConfiguration)
   ]
