+++ centreon/packages/js-config/rspack/plugins/TransformPreloadScript.js
@@ -8,7 +8,7 @@ module.exports = class TransformPreloadScript {
   }
 
   apply(compiler) {
-    compiler.hooks.done.tap('CentreonModulePlugin', () => {
+    compiler.hooks.done.tap('TransformPreloadScript', () => {
       if (!fs.existsSync(compiler.options.output.path)) {
         fs.mkdirSync(compiler.options.output.path, { recursive: true });
       }
