+++ centreon/packages/js-config/rspack/plugins/WriteRemoteEntryNameToModuleFederation.js
