+++ centreon/rspack.config.js
@@ -1,20 +1,24 @@
 const path = require('path');
 
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-const HtmlWebpackHarddiskPlugin = require('html-webpack-harddisk-plugin');
-const webpack = require('webpack');
+const rspack = require('@rspack/core');
 const { merge } = require('webpack-merge');
 
-const getBaseConfiguration = require('./packages/js-config/webpack/base');
+const getBaseConfiguration = require('./packages/js-config/rspack/base');
 
-module.exports = (jscTransformConfiguration, enableCoverage = false) =>
+module.exports = (enableCoverage = false) =>
   merge(
     getBaseConfiguration({
       enableCoverage,
-      jscTransformConfiguration,
       moduleName: 'centreon'
     }),
     {
+      devServer: {
+        devMiddleware: {
+          writeToDisk: (filename) => {
+            return /index.html$/.test(filename);
+          }
+        }
+      },
       entry: ['./www/front_src/src/index.tsx'],
       output: {
         crossOriginLoading: 'anonymous',
@@ -23,16 +27,17 @@ module.exports = (jscTransformConfiguration, enableCoverage = false) =>
         publicPath: './static/'
       },
       plugins: [
-        new webpack.ProvidePlugin({
+        new rspack.ProvidePlugin({
           React: 'react',
           process: 'process/browser'
         }),
-        new HtmlWebpackPlugin({
-          alwaysWriteToDisk: true,
+        new rspack.HtmlRspackPlugin({
           filename: path.resolve(`${__dirname}`, 'www', 'index.html'),
           template: './www/front_src/public/index.html'
         }),
-        new HtmlWebpackHarddiskPlugin()
+        new rspack.IgnorePlugin({
+          resourceRegExp: /www\/widgets/
+        })
       ],
       resolve: {
         alias: {
