+++ centreon/rspack.config.prod.js
@@ -1,0 +1,9 @@
+const { merge } = require('webpack-merge');
+
+const getBaseConfiguration = require('./rspack.config');
+
+module.exports = merge(getBaseConfiguration(), {
+  optimization: {
+    runtimeChunk: true
+  }
+});
