+++ centreon/www/widgets/rspack.config.dev.js
@@ -1,10 +1,8 @@
 const { merge } = require('webpack-merge');
 
-const {
-  getDevConfiguration
-} = require('@centreon/js-config/webpack/patch/dev');
+const { getDevConfiguration } = require('@centreon/js-config/rspack/patch/dev');
 
-const baseConfig = require('./webpack.config');
+const baseConfig = require('./rspack.config');
 
 module.exports = ({ widgetName }) =>
   merge(baseConfig({ widgetName }), getDevConfiguration());
