--- centreon/www/widgets/webpack.config.analyze.js
@@ -1,9 +1,0 @@
-const { merge } = require('webpack-merge');
-const { BundleAnalyzerPlugin } = require('webpack-bundle-analyzer');
-
-const baseConfig = require('./webpack.config');
-
-module.exports = ({ widgetName }) =>
-  merge(baseConfig({ widgetName }), {
-    plugins: [new BundleAnalyzerPlugin()]
-  });
